Factor out modules-load
This should probably stay for the long term to keep compatibility with arch's /etc/rc.conf.

Notice that this works in parallel with the native systemd version (after discussing with falconindy, I no longer believe it is worth adding arch support to the native systemd tool).

The removal of the psmouse module should probably be removed at some point.