persistent-settings: no longer set LANG in locale.sh
LANG is set by systemd and inherited by every user and every service. No need to set it when a shell is created.

Change of behaviour: if LANG is changed manually after boot, then it is no longer reset when a new shell is created. This change is intended.

Caveat: The old /etc/profile.d/locale.sh file will still be applied unless it is manually deleted.