Change criteria for setting /etc/localtime
The timestamps on the /usr/share/zoneinfo will only ever change on a
tz-data package update, so it doesn't make sense to compare the mtime of
this file with /etc/localtime.

Checking for /etc/rc.conf is also a questionable move -- it has no
direct correlation to setting the timezone other than being the source
of the TIMEZONE variable (which might be set somewhere else in the
future).

End result: check for TIMEZONE being set and the associated zoneinfo
file existing. This means we'll be setting the timezone more often than
necessary, but there's no surefire way to only set it on demand.

Signed-off-by: Dave Reisner <d@falconindy.com>