generator: use the generator hooks to generate legacy service files
This obsoletes upstream commit 980900c1d90b95e594ed0e49dcc11aa59ef14fd2. The problem with the upstream support for arch daemons was that rc.conf is strictly speaking a bash script, so the list of daemons could be any bash array. However, the code was only able to parse the most trivial kind of array.