rc.local: split rc.local and rc.local.shutdown
They are not necessarily connected, so we should run rc.local.shutdown
irrespective of whether or not rc.local succeeded. Also, we should not ignore
their return values.

We need to conditionally enable the split out unit if rc-local.service was enabled.
Adding the following to the post-upgrade script should do the trick:

  if [ -L /etc/systemd/system/multi-user.target.wants/rc-local.service ]; then

    # rc-local.service was enabled, so we should enable rc-local-shutdown.service
    ln -s /lib/systemd/system/rc-local-shutdown.service /etc/systemd/system/shutdown.target.wants/

  fi

  # reload the config after upgade
  if [ -e /sys/fs/cgroup/systemd ]; then
    systemctl daemon-reload
  fi

Signed-off-by: Tom Gundersen <teg@jklm.no>